Move dataflow var elision to its own pass
