+++ include/tvm/relax/transform.h
@@ -190,6 +190,16 @@ TVM_DLL Pass BindParams(String func_name, Map<String, runtime::NDArray> params);
 TVM_DLL Pass FoldConstant();
 
 /*!
+ * \brief If a dataflow var is used only in a binding to the dataflow block
+ * output var (i.e., a non-dataflow var), this removes the dataflow var
+ * and replaces the output var's binding with the dataflow var's direct definition.
+ *
+ * This "cleans up" a situation that commonly arises when using `CanonicalizeBindings`
+ * and `DeadCodeElimination`.
+ **/
+TVM_DLL Pass FoldDataflowBlockOutput();
+
+/*!
  * \brief Legalize high-level operator calls in Relax functions to call_tir
  * with corresponding low-level TIR PrimFuncs.
  *
