+++ python/tvm/relax/transform/transform.py
@@ -453,6 +453,21 @@ def FoldConstant() -> tvm.ir.transform.Pass:
     return _ffi_api.FoldConstant()  # type: ignore
 
 
+def FoldDataflowBlockOutput() -> tvm.ir.transform.Pass:
+    """If a dataflow var is used only in a binding to the dataflow block
+    output var (i.e., a non-dataflow var), this removes the dataflow var
+    and replaces the output var's binding with the dataflow var's direct definition.
+
+    This "cleans up" a situation that commonly arises when using `CanonicalizeBindings`
+    and `DeadCodeElimination`.
+
+    Returns
+    -------
+    ret: tvm.ir.transform.Pass
+    """
+    return _ffi_api.FoldDataflowBlockOutput()  # type: ignore
+
+
 def AnnotateTIROpPattern() -> tvm.ir.transform.Pass:
     """Annotate Op Pattern Kind for TIR functions
